% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flower_dat.R
\docType{data}
\name{flower_dat}
\alias{flower_dat}
\title{flower_dat}
\format{
A data frame with multiple columns.
}
\description{
used as demo data for this package
}
\examples{
data(flower_dat)
head(flower_dat)
}
