% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowerplot.R
\name{flowerplot}
\alias{flowerplot}
\title{flowerplot}
\usage{
flowerplot(
  data,
  angle = 90,
  a = 1,
  b = 2,
  r = 1,
  ellipse_col_pal = "Set1",
  circle_col = "white",
  label_text_cex = 1
)
}
\arguments{
\item{data}{a dataframe with multiple columns.}

\item{angle}{set the angle angle of rotation in degress.}

\item{a}{set the radii of the ellipses along the x-axes.}

\item{b}{set the radii of the ellipses along the y-axes.}

\item{r}{set the radius of the circle.}

\item{ellipse_col_pal}{set the color palette for filling the ellipse.}

\item{circle_col}{set the color for filling the circle.}

\item{label_text_cex}{set the label text cex.}
}
\value{
A flower plot.
}
\description{
A package for drawing flower plot with multiple data sets.
}
\examples{
data(flower_dat)
head(flower_dat)

flowerplot(flower_dat)
flowerplot(flower_dat, a = 0.5, b = 2, r = 1,
           circle_col = "red", ellipse_col_pal = "Spectral",
           label_text_cex = 1.5)
flowerplot(flower_dat, angle = 60, ellipse_col_pal = "Set3")
}
\author{
Wei Dong \href{mailto:dongw26@mail2.sysu.edu.cn}{dongw26@mail2.sysu.edu.cn}
}
